requirements.txt
dareplane_utils>=0.0.9
dtw>=1.4.0
fire>=0.5.0
GitPython>=3.1.43
joblib>=1.3.2
matplotlib>=3.9.2
plotly>=5.19.0
numpy==1.26.4
wx>=4.2.2
freetype-py
psychopy
pylsl>=1.16.2
pyserial>=3.5
PyYAML>=6.0.2
rich>=13.9.4
scikit_learn>=1.4.2
scipy>=1.14.1

